+++ main/cypress.config.js
@@ -8,4 +8,5 @@ module.exports = defineConfig({
   },
   viewportWidth: 1846,
   viewportHeight: 933,
+  defaultCommandTimeout: 10000,
 });
