+++ main/cypress/e2e/front.cy.js
