+++ main/cypress/e2e/searchPage.cy.js
