+++ main/package.json
@@ -10,7 +10,7 @@
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.1.1",
     "@testing-library/user-event": "^14.1.1",
-    "axios": "^1.3.2",
+    "axios": "^1.3.3",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
     "react-heart": "^1.0.6",
@@ -18,7 +18,7 @@
     "react-multi-carousel": "^2.8.2",
     "react-rating-stars-component": "^2.2.0",
     "react-router-dom": "^6.8.1",
-    "react-scripts": "5.0.1",
+    "react-scripts": "^5.0.1",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
@@ -28,7 +28,7 @@
     "eject": "react-scripts eject",
     "eslint": "eslint .",
     "cypress:open": "cypress open",
-    "test:e2e": "cypress run"
+    "test:e2e": "cypress run --browser chrome"
   },
   "eslintConfig": {
     "extends": [
