+++ main/src/components/Book.js
@@ -2,8 +2,7 @@ import axios from "axios";
 import { useState, useEffect } from "react";
 import ReactStars from "react-rating-stars-component";
 import image from "../NoImage.jpg";
-import { getCookie, setCookie, onWishlist, addToWishlist } from "../Cookies.js";
-import Heart from "react-heart";
+import { getCookie, setCookie } from "../Cookies.js";
 
 const GetBookByID = (id) => {
   const [book, setbook] = useState([]);
@@ -32,18 +31,7 @@ const Book = () => {
     isHalf: false,
     value: stars,
     onChange: (newValue) => {
-      setCookie(bookId, newValue, 5);
-    },
-  };
-  
-  var isWishlisted = onWishlist(bookId);
-
-  const heartElement = {
-    isActive: isWishlisted,
-    onClick: () => {
-      //ADDING BOOKS TO WISHLIST
-      //addToWishlist(bookId)
-      //isWishlisted = onWishlist(bookId)
+      setCookie("B", bookId, newValue, 5);
     },
   };
 
@@ -66,9 +54,6 @@ const Book = () => {
       </div>
       <h3>Your rating:</h3>
       <ReactStars {...ratingStars} />
-      <div style={{ width: "2rem"}}>
-        <Heart {...heartElement}/>
-      </div>
       <h3>Authors:</h3>
       <p>{book.authors}</p>
       <h3>Year:</h3>
