--- src/app.py
@@ -1,8 +1,0 @@
-import os
-from flask import Flask, jsonify, request
-from flask_sqlalchemy import SQLAlchemy
-
-app = Flask(__name__)
-
-
-import routes
