--- src/services/service.py
@@ -1,1 +1,0 @@
-print("HEllo World")
