--- src/static/main.css
@@ -1,12 +1,0 @@
-.body {
-    background-color: rgb(218, 227, 243);
-    color: black;
-}
-.navbar a{
-    border-style: solid;
-    border-width: 2px;
-    text-decoration: none;
-    font-size: 20px;
-    margin: 10px 10px 10px 10px;
-    padding: 5px;
-}
