--- src/templates/layout.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <link rel="stylesheet" href="/static/main.css">
-        <title>Movie Book Recommender</title>
-        <meta charset="utf-8">
-    </head>
-    <body>
-        <div class="navbar">
-            <a href="/">ItemLense</a>
-            <a href="/">My Ratings</a>
-            <a href="/">Wishlist</a>
-        </div>
-        {% block content %}{% endblock %}
-    </body>
-</html>
