--- src/templates/movies.html
@@ -1,3 +1,0 @@
-{% extends "layout.html" %}
-{% block content %}
-{% endblock %}
